Updated documentation for state_timeout
When trying to use this builder, I found I was running into timeouts when taking the final snapshot. I noticed the "stateTimeout" in the example had the wrong key name, the wrong example value, and wasn't documented.

I dug into the source, and found the config that worked, and by setting the timeout to 15m was able to get a snapshot. I've updated the documentation to reflect the correct values.